Improve tracking of modules to monitor
Replace sys.meta_path with our own import tracker while we are running
the function.  This allows us to track all attempted imports, even if
the module does not exist or if an exception occurs trying to import it.
This lets us correctly monitor all desired paths even after an error.