[cad] start converting some APIs from degrees to radians
Given that all of the stdlib math functions use radians, it is more
efficient to use radians internally.  Even though degrees are often
easier in human-facing APIs, it is better to keep the degrees-to-radians
conversion as far out towards these APIs as possible, and avoid
conversions in more internal code.